backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymupdf==1.23.7
pdfplumber==0.10.3
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
chromadb==0.4.18
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.5.2 